+++ tests/PhpWord/Tests/Style/FontTest.php
@@ -71,6 +71,9 @@ class FontTest extends \PHPUnit_Framework_TestCase
             'allCaps' => false,
             'fgColor' => null,
             'bgColor' => null,
+            'scale' => null,
+            'spacing' => null,
+            'kerning' => null,
         );
         foreach ($attributes as $key => $default) {
             $get = is_bool($default) ? "is{$key}" : "get{$key}";
@@ -106,6 +109,9 @@ class FontTest extends \PHPUnit_Framework_TestCase
             'fgColor' => Font::FGCOLOR_YELLOW,
             'bgColor' => 'FFFF00',
             'lineHeight' => 2,
+            'scale' => 150,
+            'spacing' => 240,
+            'kerning' => 10,
         );
         $object->setStyleByArray($attributes);
         foreach ($attributes as $key => $value) {